Rename flag that enables/disables _SingleLevelFunction for functorch
functorch used to have a switch that enables/disables autograd.Function.
That switch now enables/disables torch.autograd.function._SingleLevelFunction, so
I've renamed it accordingly.

We could just delete the switch because users should not be directly
working with torch.autograd.function._SingleLevelFunction. However,
it was useful for debugging when something went wrong when I was
implementing the autograd.Function <> functorch interaction, so I want
to keep it around as a debugging tool for a while since the code is
already there.

Test Plan:
- updated tests

ghstack-source-id: 990f9340dd7406c9d21bec6d40d67cd045c459d2
Pull Request resolved: https://github.com/pytorch/pytorch/pull/92025